+++ lazily/types.py
@@ -3,11 +3,20 @@ from typing import Protocol, TypeVar
 
 __all__ = ["LazilyCallable", "ResolveCallable"]
 
-C = TypeVar("C")
-T = TypeVar("T")
+C_contra = TypeVar("C_contra", contravariant=True, bound=dict)
+C_cov = TypeVar("C_cov", covariant=True, bound=dict)
+T_cov = TypeVar("T_cov", covariant=True)
 
-class LazilyCallable(Protocol[C, T]):
-    def __call__(self, ctx: C) -> T: ...
+R_contra = TypeVar("R_contra", contravariant=True)  # <-- can be ANY input type
 
-class ResolveCallable(Protocol[C]):
-    def __call__(self, ctx: C) -> dict: ...
+
+class LazilyCallable(Protocol[C_contra, T_cov]):
+    __name__: str
+
+    def __call__(self, ctx: C_contra) -> T_cov: ...
+
+
+class ResolveCallable(Protocol[R_contra, C_cov]):
+    __name__: str
+
+    def __call__(self, ctx: R_contra) -> C_cov: ...