+++ CHANGELOG.md
@@ -1,12 +1,19 @@
+## 0.9.0
+
+* Add type-aware context to slot and cell functions.
+* The ctx argument type can be enforced via mypy.
+* mypy tests now pass
+
 ## 0.8.0
 
 * Add Cell.get() as an alias to Cell.value
-* Add Cell.set(value) as an alias to Cell.value = value 
+* Add Cell.set(value) as an alias to Cell.value = value
 
 ## 0.7.0
 
 * Add @slot_def and @cell_def decorators...which both take a resolve_ctx callable to resolve the ctx object.
-* Allows for calling a slot function with an object that contains the ctx, such as a Request or Graphene GraphQLResolveInfo.
+* Allows for calling a slot function with an object that contains the ctx, such as a Request or Graphene
+  GraphQLResolveInfo.
 
 ## 0.6.0